Check HTTP response status code
We used to just pass through the raven-clj http request response without
checking for success. We now check for status == 200 (Sentry seems to
return 200 for success and 400 for errors), and we parse Sentry's error
report ID from the response. The caller should probably print out the ID
to make it easier to correlate logs and error reports.

Fixes #5