Capture individual exceptions
Instead of capturing just Throwable, we now capture also NPEs and
JsonGenerationExceptions. The former should only be caused by the DSN
being of an incorrect format to the parsing fails, and the latter is
caused when raven-clj tries to serialize the event to JSON and some
fields are not serializable.

Fixes #6