The next release will be 0.3.0
- Assert that `extra` argument is a maps

Fixes #16

- Introduce an ErrorReporter protocol

We introduce an ErrorReporter protocol and a CoraxErrorReporter
record. The record wraps the DSN and log fn that were previously passed
to `report`.

Fixes #17

- Introduce a Logger protocol

We used to have a log function which was passed in an assortment of
arguments and the callee needed to work out what was happening. Now we
have a protocol with a method for each type of error-report result.

Fixes #18